backend/requirements.txt
fastapi
uvicorn
PyPDF2
pymupdf
python-multipart
pillow
pdfkit